refactor(create-dialog): renames Dialog.js and moves into Exercises directory
