+++ src/Components/Layouts/Header.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppBar, Toolbar, Typography } from '@material-ui/core';
-import CreateDialog from '../Exercises/Dialogs/Create';
+import CreateDialog from '../Exercises/CreateDialog';
 
 export default ({ muscles, createExercise }) =>
     <AppBar position="static">
